pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmscan"
dynamic = ["version"]
description = "A fast, modern CLI that shows which local LLMs fit on your machine"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{ name = "LLM-Check Contributors" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: System :: Hardware",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "typer>=0.12.3,<1.0",
  "rich>=13.7.1,<14.0",
  "httpx>=0.27,<1.0",
]

[project.urls]
Homepage = "https://github.com/adityaarakeri/llmscan"
Repository = "https://github.com/adityaarakeri/llmscan"
Issues = "https://github.com/adityaarakeri/llmscan/issues"

[project.scripts]
llmscan = "llmscan.cli:app"
llmc = "llmscan.cli:app"

[project.optional-dependencies]
test = ["pytest>=8.0,<9.0", "pyyaml>=6.0,<7.0", "tomli>=2.0,<3.0; python_version < '3.11'"]
dev = ["ruff>=0.4,<1.0", "mypy>=1.10,<2.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.setuptools.dynamic]
version = {attr = "llmscan.__version__"}

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
llmscan = ["models.json"]

[tool.setuptools.packages.find]
where = ["."]
include = ["llmscan*"]
